Reorganize test helpers
