+++ ui/e2e/utils/files.ts
@@ -1,5 +1,6 @@
 import { CreateFileSchemaType } from "~/api/files/schema";
 import { createFile as apiCreateFile } from "~/api/files/mutate/createFile";
+import { getFile as apiGetFile } from "~/api/files/query/file";
 import { encode } from "js-base64";
 import { headers } from "./testutils";
 
@@ -52,3 +53,39 @@ export const createDirectory = async ({
     },
     headers,
   });
+
+type ErrorType = { json?: { code?: string } };
+
+export const checkIfFileExists = async ({
+  namespace,
+  path,
+}: {
+  namespace: string;
+  path: string;
+}) => {
+  try {
+    const response = await apiGetFile({
+      urlParams: {
+        baseUrl: process.env.PLAYWRIGHT_UI_BASE_URL,
+        path,
+        namespace,
+      },
+    });
+
+    if (!response.data) {
+      throw `Fetching file at ${path} in namespace ${namespace} failed`;
+    }
+
+    return response.data.path === path;
+  } catch (error) {
+    const typedError = error as ErrorType;
+
+    if (typedError?.json?.code === "resource_not_found") {
+      return false;
+    }
+
+    throw new Error(
+      `Unexpected error fetching ${path} in namespace ${namespace}`
+    );
+  }
+};
